requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
branca==0.8.1
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
chromedriver-py==138.0.7204.92
click==8.1.8
colorama==0.4.6
comtypes==1.4.11
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
edge-tts==7.0.2
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.18.0
Flask==3.1.0
folium==0.20.0
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
geocoder==1.38.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
gTTS==2.5.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.33.2
HyperPyYAML==1.2.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
Js2Py==0.74
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mistune==3.1.4
MouseInfo==0.1.3
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.5
noisereduce==3.0.3
numba==0.61.2
numpy==2.2.6
openai==1.97.0
opencv-python==4.12.0.88
outcome==1.3.0.post0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.5
pefile==2023.2.7
pickleshare==0.7.5
pillow==11.3.0
pipreqs==0.5.0
pipwin==0.5.2
platformdirs==4.3.8
playsound==1.2.2
playsound3==3.2.4
pooch==1.8.2
proglog==0.1.12
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycaw==20240210
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.2
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
pyjsparser==2.7.1
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
pypiwin32==223
PyPrind==2.11.3
PyRect==0.2.0
PyScreeze==1.0.1
pySmartDL==1.3.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyttsx3==2.98
pytweening==1.2.0
pywhatkit==5.4
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==27.1.0
ratelim==0.1.6
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-learn==1.7.0
scipy==1.16.0
screen_brightness_control==0.24.2
selenium==4.35.0
sentencepiece==0.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
speechbrain==1.0.3
SpeechRecognition==3.14.3
srt==3.5.3
stack-data==0.6.3
stem==1.8.2
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.6.0
tinycss2==1.4.0
torch==2.8.0
torchaudio==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
vosk==0.3.45
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wikipedia==1.4.0
winshell==0.6
WMI==1.5.1
wsproto==1.2.0
xyzservices==2025.4.0
yarg==0.1.9
yarl==1.20.1
